@@ -5,9 +5,9 @@ import 'package:uuid/uuid.dart';
 
 const MethodChannel _channel = const MethodChannel('ssh');
 const EventChannel _eventChannel = const EventChannel('shell_sftp');
-Stream<dynamic> _onStateChanged;
+Stream<dynamic>? _onStateChanged;
 
-Stream<dynamic> get onStateChanged {
+Stream<dynamic>? get onStateChanged {
   if (_onStateChanged == null) {
     _onStateChanged =
         _eventChannel.receiveBroadcastStream().map((dynamic event) => event);
@@ -18,26 +18,25 @@ Stream<dynamic> get onStateChanged {
 typedef void Callback(dynamic result);
 
 class SSHClient {
-  String id;
+  late String id;
   String host;
   int port;
   String username;
   dynamic passwordOrKey;
-  StreamSubscription<dynamic> stateSubscription;
-  Callback shellCallback;
-  Callback uploadCallback;
-  Callback downloadCallback;
+  StreamSubscription<dynamic>? stateSubscription;
+  Callback? shellCallback;
+  Callback? uploadCallback;
+  Callback? downloadCallback;
 
   SSHClient({
-    @required this.host,
-    @required this.port,
-    @required this.username,
-    @required
-        this.passwordOrKey, // password or {privateKey: value, [publicKey: value, passphrase: value]}
+    required this.host,
+    required this.port,
+    required this.username,
+    required this.passwordOrKey, // password or {privateKey: value, [publicKey: value, passphrase: value]}
   }) {
     var uuid = new Uuid();
     id = uuid.v4();
-    stateSubscription = onStateChanged.listen((dynamic result) {
+    stateSubscription = onStateChanged?.listen((dynamic result) {
       _parseOutput(result);
     });
   }
@@ -46,15 +45,15 @@ class SSHClient {
     switch (result["name"]) {
       case "Shell":
         if (shellCallback != null && result["key"] == id)
-          shellCallback(result["value"]);
+          shellCallback?.call(result["value"]);
         break;
       case "DownloadProgress":
         if (downloadCallback != null && result["key"] == id)
-          downloadCallback(result["value"]);
+          downloadCallback?.call(result["value"]);
         break;
       case "UploadProgress":
         if (uploadCallback != null && result["key"] == id)
-          uploadCallback(result["value"]);
+          uploadCallback?.call(result["value"]);
         break;
     }
   }
@@ -77,20 +76,16 @@ class SSHClient {
     });
     return result;
   }
-  
+
   Future<String> portForwardL(int rport, int lport, String rhost) async {
-    var result = await _channel.invokeMethod('portForwardL', {
-      "id": id,
-      "rhost": rhost,
-      "rport": rport,
-      "lport": lport
-    });
+    var result = await _channel.invokeMethod('portForwardL',
+        {"id": id, "rhost": rhost, "rport": rport, "lport": lport});
     return result;
   }
 
   Future<String> startShell({
     String ptyType = "vanilla", // vanilla, vt100, vt102, vt220, ansi, xterm
-    Callback callback,
+    Callback? callback,
   }) async {
     shellCallback = callback;
     var result = await _channel.invokeMethod('startShell', {
@@ -131,8 +126,8 @@ class SSHClient {
   }
 
   Future<String> sftpRename({
-    @required String oldPath,
-    @required String newPath,
+    required String oldPath,
+    required String newPath,
   }) async {
     var result = await _channel.invokeMethod('sftpRename', {
       "id": id,
@@ -167,9 +162,9 @@ class SSHClient {
   }
 
   Future<String> sftpDownload({
-    @required String path,
-    @required String toPath,
-    Callback callback,
+    required String path,
+    required String toPath,
+    Callback? callback,
   }) async {
     downloadCallback = callback;
     var result = await _channel.invokeMethod('sftpDownload', {
@@ -187,9 +182,9 @@ class SSHClient {
   }
 
   Future<String> sftpUpload({
-    @required String path,
-    @required String toPath,
-    Callback callback,
+    required String path,
+    required String toPath,
+    Callback? callback,
   }) async {
     uploadCallback = callback;
     var result = await _channel.invokeMethod('sftpUpload', {
@@ -218,16 +213,19 @@ class SSHClient {
     shellCallback = null;
     uploadCallback = null;
     downloadCallback = null;
-    stateSubscription.cancel();
+    stateSubscription?.cancel();
     _channel.invokeMethod('disconnect', {
       "id": id,
     });
   }
-  
-   Future<bool> isConnected() async {
+
+  Future<bool> isConnected() async {
     bool connected = false; // default to false
-    var result =  await _channel.invokeMethod('isConnected', {"id": id,});
-    if (result == "true") { // results returns a string, therefor we need to check the string 'true'
+    var result = await _channel.invokeMethod('isConnected', {
+      "id": id,
+    });
+    if (result == "true") {
+      // results returns a string, therefor we need to check the string 'true'
       connected = true;
     }
     return connected;
